+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [PEP 440 about versioning](https://www.python.org/dev/pe
 - Allow using spaces around colon when naming a resource in a query.
 - Better exceptions
 - Manage slicing/list indexing in filters
+- Change in the format for lists, allowing to return list of values, not only list of dicts
 
 ## [0.0.1a1] - 2015-06-15
 - First working version